+++ Source/Events.Store.MongoDB/EventStore.cs
@@ -58,7 +58,7 @@ namespace Dolittle.Runtime.Events.Store.MongoDB
                 return await session.WithTransactionAsync(
                     async (transaction, cancel) =>
                     {
-                        var eventLogSequenceNumber = (uint)await _connection.EventLog.CountDocumentsAsync(
+                        var eventLogSequenceNumber = (ulong)await _connection.EventLog.CountDocumentsAsync(
                             transaction,
                             _eventFilter.Empty,
                             cancellationToken: cancel).ConfigureAwait(false);
@@ -97,7 +97,7 @@ namespace Dolittle.Runtime.Events.Store.MongoDB
                 return await session.WithTransactionAsync(
                     async (transaction, cancel) =>
                     {
-                        var eventLogSequenceNumber = (uint)await _connection.EventLog.CountDocumentsAsync(
+                        var eventLogSequenceNumber = (ulong)await _connection.EventLog.CountDocumentsAsync(
                             transaction,
                             _eventFilter.Empty,
                             cancellationToken: cancel).ConfigureAwait(false);
