+++ Source/Events.Store/AggregateRootVersion.cs
@@ -8,7 +8,7 @@ namespace Dolittle.Runtime.Events.Store
     /// <summary>
     /// Represents a version of an aggregate root as a natural number, corresponding to the number of events the Aggregate Root has applied to an Event Source.
     /// </summary>
-    public class AggregateRootVersion : ConceptAs<uint>
+    public class AggregateRootVersion : ConceptAs<ulong>
     {
         /// <summary>
         /// The initial version of an aggregate root that has applied no events.
@@ -16,9 +16,9 @@ namespace Dolittle.Runtime.Events.Store
         public static readonly AggregateRootVersion Initial = 0;
 
         /// <summary>
-        /// Implicitly convert a <see cref="uint"/> to an <see cref="AggregateRootVersion"/>.
+        /// Implicitly convert a <see cref="ulong"/> to an <see cref="AggregateRootVersion"/>.
         /// </summary>
         /// <param name="number">The number.</param>
-        public static implicit operator AggregateRootVersion(uint number) => new AggregateRootVersion { Value = number };
+        public static implicit operator AggregateRootVersion(ulong number) => new AggregateRootVersion { Value = number };
     }
 }
